move address search box to header
